server/requirements.txt
uvicorn[standard]==0.15.0
gunicorn==20.1.0
fastapi==0.68.1
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiofiles==0.7.0
python-dotenv
requests
shapely
rasterio
matplotlib